Bestellingen fe backup

See merge request root/Case03!59

+++ bestellingservice/src/main/java/nl/kantilever/bestellingservice/entities/Gebruiker.java
@@ -33,9 +33,9 @@ public class Gebruiker {
 
   private String telefoonnummer;
 
-  private int maxKrediet = 500; // default value
+  private double maxKrediet = 500.0; // default value
 
-  private int huidigKrediet = 0; // default value
+  private double huidigKrediet = 0.0; // default value
 
 
   public Gebruiker(long id, String voornaam, String achternaam) {
@@ -136,19 +136,19 @@ public class Gebruiker {
     this.telefoonnummer = telefoonnummer;
   }
 
-  public int getMaxKrediet() {
+  public double getMaxKrediet() {
     return maxKrediet;
   }
 
-  public void setMaxKrediet(int maxKrediet) {
+  public void setMaxKrediet(double maxKrediet) {
     this.maxKrediet = maxKrediet;
   }
 
-  public int getHuidigKrediet() {
+  public double getHuidigKrediet() {
     return huidigKrediet;
   }
 
-  public void setHuidigKrediet(int huidigKrediet) {
+  public void setHuidigKrediet(double huidigKrediet) {
     this.huidigKrediet = huidigKrediet;
   }
